refactor(utils): Split utils.lua into separate modules (os, path)
