+++ lua/im-switch/utils/init.lua
@@ -1,0 +1,4 @@
+return {
+  os = require("im-switch.utils.os"),
+  path = require("im-switch.utils.path"),
+}
